client: add state mod
ExptesslaneState and DeviceNetworkState are also part of interfaces.
Such that we put state.rs a mod.
